requirements.txt
# Core AI and ML libraries
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.8
langchain-core>=0.1.0
google-generativeai>=0.3.0

# Vector database and embeddings
faiss-cpu>=1.7.4

# Voice/Audio features
gTTS>=2.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0